feat: not delete or edit a rented book

+++ src/services/books.service.ts
@@ -1,12 +1,16 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { CreateBookDTO } from 'src/dtos/create-book.dto';
 import { UpdateBookDTO } from 'src/dtos/update-book.dto';
 import { IBook } from 'src/interfaces/book.interface';
 import { BooksRepository } from 'src/repositories/books.repository';
+import { RentsRepository } from 'src/repositories/rents.repository';
 
 @Injectable()
 export class BooksService {
-  constructor(private booksRepository: BooksRepository) { }
+  constructor(
+    private booksRepository: BooksRepository,
+    private readonly rentsRepository: RentsRepository,
+  ) { }
 
   async create(createBookDTO: CreateBookDTO): Promise<IBook> {
     return await this.booksRepository.create(createBookDTO);
@@ -17,10 +21,22 @@ export class BooksService {
   }
 
   async update(id: string, updateBookDTO: UpdateBookDTO): Promise<IBook> {
+    const bookIsRented = await this.rentsRepository.checkIfTheBookIsRented(id);
+
+    if (bookIsRented) {
+      throw new BadRequestException();
+    }
+
     return await this.booksRepository.update(id, updateBookDTO);
   }
 
   async delete(id: string): Promise<IBook> {
+    const bookIsRented = await this.rentsRepository.checkIfTheBookIsRented(id);
+
+    if (bookIsRented) {
+      throw new BadRequestException();
+    }
+
     return await this.booksRepository.delete(id);
   }
